[PORT FROM R1] Video: ww40.5 Porting local psbvideo to MFLD Android
BZ: 11126, 11008, 11086, 11496

Delay overlay init to the first render and add more error handling when create overlay cmd buffer.

Origianlly when create decode surfaces, the surface is registered into BCD
But the surface may not be used for rendering. This patch delay surface
BCD register later, so that only when rendering, the surface is registered into BCD

Change-Id: Id2250e9096a6e64aca322c0c85d496efc4998d94
Orig-Change-Id: Ie4dc9bea3797bc06cfed54816fb0f9afaa9f8863
Signed-off-by: Kun Wang <kun.k.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/20863
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: Yuan, Shengquan <shengquan.yuan@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>