assign the stride with graphic->stride instead of width
BZ: 44326

On the widi mode the surface was create by widi in the use space, the
stride was assigned as the video width.  However after the video enhancement
implemented, the camera source transmit the data which has padding to
the encoder, make the width different with the stride.
So modify the stride assignment to fix the display problem.
Signed-off-by: Ji Guoliang <guoliang.ji@gmail.com>

Change-Id: Ie70c89fc56f180b289266de58bd9c0bb53d34037
Reviewed-on: http://android.intel.com:8080/54829
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>