Set minqp to 2 in H263 encoding.
BZ: 55358

The particular macroblocks this occurs on have a sharp white to black transition,
and the coefficients generated by the forward DCT on intra blocks are large.
With the QP of 1 being used, after quantization the first AC coefficient is
still larger then H263 can encode. So the h/w is clipping the coefficient to
the maximum value allowed by H263. Specifically, the first AC coeff is 227
after quantization, but is being clipped to range +/-127. That's why visual
artifacts are seen. Set minqp to 2 as WR. IMG will continue to look into this
issue and find a way to fix it in encoder firmware.

Change-Id: If4f057a52e298c2224e9f37e1a49faa1aeecb0dd
Signed-off-by: Elaine Wang <elaine.wang@intel.com>
Signed-off-by: hding3 <haitao.ding@intel.com>
Reviewed-on: http://android.intel.com:8080/66640
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>