[PORT FROM MAIN] Encoder: Cope with wrong target bitrate
BZ: 67609

The exported video is messy when trimming the beginning and end of the video.
This bug can be reproduce when use 50M as target bitrate and 5 as frame rate.
It violates ITU-T Rec. H.264 (03/2005) A.3.1. Video driver didn't cope with
this case correctly and caused wrong initial QP. To fix this, video driver
force the bitrate to be under the limitation.
Signed-off-by: Elaine Wang <elaine.wang@intel.com>

Change-Id: If0b906544edc78ada54a26c6049c3c07f30cc41d
Reviewed-on: http://android.intel.com:8080/80443
Reviewed-by: Zhao, Leo <leo.zhao@intel.com>
Reviewed-by: Ding, Haitao <haitao.ding@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>