Added input source surface stride's check for 64-bytes alignment into MERR JPEG Encode Driver.
BZ: 103234

Issue: No checking and the address of a wrong source
data buffer was sent to HW to encode anyway, in which
case encoding failed with timeout errors and the driver
never told its users what their problems really were.

Solution: Added checking code and did not let encoding
go on if alignment was not satisfied.

Change-Id: I4f84ee183bc0fcf1f42fd778e780c7f93b16558a
Signed-off-by: SUN,Jing <jing.a.sun@intel.com>
Reviewed-on: http://android.intel.com:8080/104571
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>