[PORT FROM R42S]Encoder: Support for dynamic Min QP change
BZ: 116990

It's a request from TW. They need libva to support dynamic
Min QP change. Then WiDi can set a bigger min QP for I frame to
reduce the I frame size.
Note: The min QP can't be zero if it's changing dyanmically.
Also correct the cpb intial level and delay when window size
is changed.
Refined the RC parameter calculation in VCM mode. With this change,
the resulted average bitrate will be closed to target one. It used to
be 90% of target bitate.

Change-Id: I441deebecd8a9e95dbdb5fb0a2a16e6487743c3c
Signed-off-by: Elaine Wang <elaine.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/115496
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>