Encoder: Support for dynamic Min QP change
BZ: 116990

It's a request from TW. They need libva to support dynamic
Min QP change. Then WiDi can set a bigger min QP for I frame to
reduce the I frame size.
Note: The min QP can't be zero if it's changing dyanmically.
Also correct the cpb intial level and delay when window size
is changed.
Refined the RC parameter calculation in VCM mode. With this change,
the resulted average bitrate will be closed to target one. It used to
be 90% of target bitate.

Change-Id: I088d64e62328a414e2b617e5555c27adb468fd33
Signed-off-by: Elaine Wang <elaine.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/115003
Reviewed-by: Yuan, Shengquan <shengquan.yuan@intel.com>
Tested-by: Ding, Haitao <haitao.ding@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>