Fixed WidiComposer video background setting issue.
To compose RGB & Video together, composer needs to know which layer is bigger.
The bigger one would be taken as the container layer, which is the background.
The judgement is wrong in user space driver, it cause when composing HD content
together, smaller video layer is taken as background by mistake.
This patch fixed this issue.

Change-Id: I426be9d475750bc9cd24d70689fd2b3ac13fbdc7
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-5703
Signed-off-by: Wang Kun <kun.k.wang@intel.com>
Reviewed-on: https://android.intel.com:443/289747