psb_video: clear video buffer that from gralloc to be black
For some clip, only one frame is key frame, that is the first frame.
Theoretically, this kind of clip cannot support seek. (We know, after
suspend/resume, seek will be performed). But for some P frames within
these clips, most of MBs (except for the MBs located at the most bottom
row and the most right colomn) are intra-coded, so we can see these MBs
can be correctly displayed and the other ones are filled with garbage.

The solution is to clear buffer to be black before decoding one frame.

Change-Id: I109920e0aead07bb75e26157a84116f6513b8264
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-20253
Signed-off-by: Tianmi Chen <tianmi.chen@intel.com>
Reviewed-on: https://android.intel.com:443/330437